+++ src/Core/MainMenu/Systems/IsHovering.hpp
@@ -15,7 +15,7 @@ namespace Arcade::Game
         IsHovering();
         ~IsHovering() override = default;
 
-        void run(float deltaTime, Arcade::ECS::IEventManager &eventManager,
+        void run(double deltaTime, Arcade::ECS::IEventManager &eventManager,
             Arcade::ECS::IEntityManager &currentEntityManager) override;
     };
 } // namespace Arcade::Game
