+++ src/Game/libs/TestGame/TestGame.hpp
@@ -19,7 +19,7 @@ namespace Arcade::Game
         TestGame();
         ~TestGame() override = default;
 
-        void update(float deltaTime, Arcade::ECS::IEventManager &eventManager) override;
+        void update(double deltaTime, Arcade::ECS::IEventManager &eventManager) override;
         void handleEventSceneChange(Arcade::ECS::IEventManager &eventManager) override;
     };
 } // namespace Arcade::Game
